requirements.txt
flask
requests
psycopg2-binary
python-dotenv
gunicorn
pytz
google-generativeai